moved EchoMessage into admin module
