+++ backend/src/messages/admin.py
@@ -40,3 +40,18 @@ class LogMessage(Message):
         text = self.message.get(MessageAttribute.WS_ATTR_MESSAGE)
         caller = self.message.get(MessageAttribute.WS_ATTR_CALLER)
         getLogger(caller or "FrontEnd").log(LOGGING_LEVEL.get(level), text)
+
+
+class EchoMessage(Message):
+    "Echo request from Frontend. Reply by sending the requested message back."
+
+    @classmethod
+    def message_type(cls):
+        return MessageType.WS_TYPE_ECHO
+
+    async def handle_message(self, connection: "WS_Connection"):
+        "Return the payload to the requsted component"
+        await connection.send_message_to_component(
+            self.message.get(MessageAttribute.WS_ATTR_COMPONENT),
+            self.message.get(MessageAttribute.WS_ATTR_PAYLOAD),
+        )
